Functions in R Programming
Functions are useful when you want to perform a certain task multiple times. A function accepts input arguments and produces the output by executing valid R commands that are inside the function. In R Programming Language when you are creating a function the function name and the file in which you are creating the function need not be the same and you can have one or more function definitions in a single R file.
Types of function in R Language

    Built-in Function: Built function R is sq(), mean(), max(), these function are directly call in the program by users.
    User-defined Function: R language allow us to write our own function.