R Prog. Update by adding Excercise
R is a language and environment for statistical computing and graphics. It is a GNU project which is similar to the S language and environment which was developed at Bell Laboratories by John Chambers and colleagues. R can be considered as a different implementation of S. R provides a wide variety of statistical and graphical techniques, and is highly extensible.